refactor(macos-settings): simplify general tab and consolidate developer tooling
